Restructure files and add cursor image
