+++ client/app/bundles/Room/components/ActionBox/PointBar/PointBar.jsx
@@ -1,6 +1,7 @@
 import PropTypes from 'prop-types'
 import React from 'react'
 import BarColors from 'libs/barColors'
+import CursorImage from './check.png'
 
 export default class PointBar extends React.Component {
   selectPoint = () => {
@@ -14,15 +15,17 @@ export default class PointBar extends React.Component {
 
   render() {
     var pointIndicator = BarColors.emoji(this.props.point) || this.props.point;
+    let cursorStyle
+    if ("Moderator" === this.props.role) {
+      cursorStyle = { cursor: `url(${CursorImage}), auto` }
+    }
 
     return (
       <li className="row" data-point={this.props.point}>
-        <div className="row-container">
-          <div className="col-md-2 point">{pointIndicator}</div>
-          <div className="col-md-9 bar">
-            <div onClick={this.selectPoint} style={{width: this.props.barWidth + '%', background: this.props.color, color: '#fff', 'textAlign': 'center'}}>
-              {this.props.count}
-            </div>
+        <div className="col-md-2 point">{pointIndicator}</div>
+        <div className="col-md-9 bar" style={cursorStyle}>
+          <div onClick={this.selectPoint} style={{width: this.props.barWidth + '%', background: this.props.color, color: '#fff', 'textAlign': 'center'}}>
+            {this.props.count}
           </div>
         </div>
       </li>
