+++ client/app/bundles/Room/containers/Room.jsx
@@ -4,7 +4,7 @@ import StatusBar from '../components/StatusBar'
 import VoteBox from '../components/VoteBox'
 import StoryListBox from '../components/StoryListBox'
 import PeopleListBox from '../components/PeopleListBox'
-import ActionBox from '../components/ActionBox'
+import ActionBox from '../components/ActionBox/ActionBox'
 import Board from '../components/Board'
 import ActionCable from 'libs/actionCable'
 import update from 'immutability-helper'
